Move getApartments logic into actions.ts
