+++ src/app/apartments/page.tsx
@@ -1,38 +1,9 @@
 import Link from 'next/link'
-import { createClient } from '@/utils/supabase/server'
-import { QueryData } from '@supabase/supabase-js'
-import { redirect } from 'next/navigation'
 import Image from 'next/image'
+import { getApartments } from './actions'
 
 export default async function Apartments() {
-  const supabase = createClient()
-
-  const { data: authData, error: authError } = await supabase.auth.getUser()
-  if (authError || !authData?.user) {
-    redirect('/login')
-  }
-
-  const apartmentsQuery = supabase
-    .from("apartments")
-    .select(`
-      id,
-      created_at,
-      name,
-      location,
-      price,
-      currency,
-      description,
-      rooms: rooms(*)
-        `
-    )
-    .order('created_at', { referencedTable: 'rooms', ascending: true })
-    .eq('user_id', authData.user.id);
-  type Apartments = QueryData<typeof apartmentsQuery>;
-
-  const { data, error } = await apartmentsQuery;
-  if (error) throw error;
-  console.log(data);
-  const apartments: Apartments = data;
+  const apartments = await getApartments();
 
   return (
     <main className="container mx-auto flex flex-col grow items-center">
